+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- File types for Ruby, F#, Shell scripts, Batch scripts, SQL, Visual Basic, Prisma schema, HTTP request files, license files
+
+### Changed
+
+- Redesigned icons for Java, JSON, XML, SQL, Visual Basic, Image files 
+- Updated KPresence to 0.6.2
+
 ## [1.3.2] - 2024-05-21
 
 ### Fixed
@@ -105,8 +114,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release
 
-[Unreleased]: https://github.com/pandier/intellij-discord-rp/compare/v1.3.2...HEAD
-[1.3.2]: https://github.com/pandier/intellij-discord-rp/compare/v1.3.1...v1.3.2
+[Unreleased]: https://github.com/pandier/intellij-discord-rp/compare/v1.3.1...HEAD
 [1.3.1]: https://github.com/pandier/intellij-discord-rp/compare/v1.3.0...v1.3.1
 [1.3.0]: https://github.com/pandier/intellij-discord-rp/compare/v1.2.1...v1.3.0
 [1.2.1]: https://github.com/pandier/intellij-discord-rp/compare/v1.2.0...v1.2.1